changed to float types

+++ db/migrate/20220414183851_add_latitude_to_stations.rb
@@ -1,5 +1,5 @@
 class AddLatitudeToStations < ActiveRecord::Migration[6.1]
   def change
-    add_column :stations, :latitude, :integer
+    add_column :stations, :latitude, :float
   end
 end
