+++ db/migrate/20220414183932_add_longitude_to_stations.rb
@@ -1,5 +1,5 @@
 class AddLongitudeToStations < ActiveRecord::Migration[6.1]
   def change
-    add_column :stations, :longitude, :integer
+    add_column :stations, :longitude, :float
   end
 end
